api/requirements.txt
azure-core==1.30.1
azure-storage-blob==12.19.1
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.2.2
cryptography==42.0.5
debugpy==1.8.1
decorator==5.1.1
distro==1.9.0
et-xmlfile==1.1.0
executing==2.0.1
fastapi==0.110.0
ipython==8.22.2
isodate==0.6.1
matplotlib-inline==0.1.6
nest-asyncio==1.6.0
numpy==1.26.4
openai==1.3.9
openpyxl==3.1.2
packaging==24.0
pandas==2.2.1
parso==0.8.3
platformdirs==4.2.0
prompt-toolkit==3.0.43
psutil==5.9.8
pure-eval==0.2.2
pycparser==2.21
pydantic==2.6.4
pydantic_core==2.16.3
Pygments==2.17.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
azure-communication-email
uvicorn==0.28.0